% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdxj_html_format.R
\name{ssdxj_format}
\alias{ssdxj_format}
\title{Custom HTML template}
\usage{
ssdxj_format(toc = TRUE, ...)
}
\arguments{
\item{toc}{should a table of contents be displayed?}

\item{...}{additional arguments provided to \@code{html_document}}
}
\description{
Loads additional style and template file
}
